🔨 [util] Move function `updatetemplatevariable` to `tests.util.variables`
